+++ include/vrv/iomei.h
@@ -82,6 +82,7 @@ class Mdiv;
 class Measure;
 class Mensur;
 class MeterSig;
+class MeterSigGrp;
 class MNum;
 class Mordent;
 class MRest;
@@ -255,6 +256,7 @@ private:
     void WriteTuning(pugi::xml_node currentNode, Tuning *tuning);
     void WriteCourse(pugi::xml_node currentNode, Course *course);
     void WriteMeasure(pugi::xml_node currentNode, Measure *measure);
+    void WriteMeterSigGrp(pugi::xml_node currentNode, MeterSigGrp *meterSigGrp);
     void WriteFb(pugi::xml_node currentNode, Fb *fb);
     void WriteStaff(pugi::xml_node currentNode, Staff *staff);
     void WriteLayer(pugi::xml_node currentNode, Layer *layer);
@@ -513,6 +515,8 @@ private:
     bool ReadCourse(Object *parent, pugi::xml_node course);
     bool ReadMeasure(Object *parent, pugi::xml_node measure);
     bool ReadMeasureChildren(Object *parent, pugi::xml_node parentNode);
+    bool ReadMeterSigGrp(Object *parent, pugi::xml_node meterSigGrp);
+    bool ReadMeterSigGrpChildren(Object *parent, pugi::xml_node parentNode);
     bool ReadFb(Object *parent, pugi::xml_node fb);
     bool ReadFbChildren(Object *parent, pugi::xml_node parentNode);
     bool ReadNum(Object *parent, pugi::xml_node parentNode);