+++ include/vrv/slur.h
@@ -67,8 +67,8 @@ public:
     /**
      * Calculate slur left/right maximum shifts required for slur not to overlap with other objects
      */
-    std::pair<int, int> CalculateAdjustedSlurShift(FloatingCurvePositioner *curve, const BezierCurve &bezierCurve, int margin,
-        bool forceBothSides, bool &isNotAdjustable);
+    std::pair<int, int> CalculateAdjustedSlurShift(FloatingCurvePositioner *curve, const BezierCurve &bezierCurve,
+        int margin, bool forceBothSides, bool &isNotAdjustable);
 
     float GetAdjustedSlurAngle(Doc *doc, Point &p1, Point &p2, curvature_CURVEDIR curveDir, bool withPoints);
     void GetControlPoints(BezierCurve &curve, curvature_CURVEDIR curveDir, bool ignoreAngle = false);