+++ src/layerelement.cpp
@@ -36,6 +36,7 @@
 #include "measure.h"
 #include "mensur.h"
 #include "metersig.h"
+#include "metersiggrp.h"
 #include "mrest.h"
 #include "mrpt.h"
 #include "mrpt2.h"
@@ -947,6 +948,7 @@ int LayerElement::AlignHorizontally(FunctorParams *functorParams)
     Note *noteParent = dynamic_cast<Note *>(this->GetFirstAncestor(NOTE, MAX_NOTE_DEPTH));
     Rest *restParent = dynamic_cast<Rest *>(this->GetFirstAncestor(REST, MAX_NOTE_DEPTH));
     TabGrp *tabGrpParent = dynamic_cast<TabGrp *>(this->GetFirstAncestor(TABGRP, MAX_TABGRP_DEPTH));
+    const bool ligatureAsBracket = params->m_doc->GetOptions()->m_ligatureAsBracket.GetValue();
 
     if (chordParent) {
         m_alignment = chordParent->GetAlignment();
@@ -963,7 +965,7 @@ int LayerElement::AlignHorizontally(FunctorParams *functorParams)
     else if (this->Is({ DOTS, FLAG, STEM })) {
         assert(false);
     }
-    else if (ligatureParent && this->Is(NOTE)) {
+    else if (ligatureParent && this->Is(NOTE) && !ligatureAsBracket) {
         // Ligature notes are all aligned with the first note
         Note *note = vrv_cast<Note *>(this);
         assert(note);
@@ -1025,6 +1027,8 @@ int LayerElement::AlignHorizontally(FunctorParams *functorParams)
             type = ALIGNMENT_SCOREDEF_METERSIG;
         else if (this->GetScoreDefRole() == SCOREDEF_CAUTIONARY)
             type = ALIGNMENT_SCOREDEF_CAUTION_METERSIG;
+        else if (this->GetParent() && this->GetParent()->Is(METERSIGGRP))
+            type = ALIGNMENT_SCOREDEF_METERSIG;
         else {
             // replace the current meter signature
             params->m_currentMeterSig = vrv_cast<MeterSig *>(this);
@@ -1783,7 +1787,8 @@ int LayerElement::AdjustXPos(FunctorParams *functorParams)
                 bool hasOverlap = this->HorizontalContentOverlap(boundingBox, margin);
 
                 if (hasOverlap) {
-                    // For note to note alignment, make sure there is a standard spacing even if they to not overlap vertically
+                    // For note to note alignment, make sure there is a standard spacing even if they to not overlap
+                    // vertically
                     if (this->Is(NOTE) and element->Is(NOTE)) {
                         overlap = std::max(overlap, element->GetSelfRight() - this->GetSelfLeft() + margin);
                     }