+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Improved rendering to prevent rendering barlines for the invisible measure (@eNote-GmbH)
 * Improved rendering of octave brackets ending (@eNote-GmbH)
 * Option --breaks-no-widow to prevent a single measure on last page (@eNote-GmbH)
+* Option --ligature-as-bracket to render ligatures as bracket instead of original notation
 * Build option NO_MXL_SUPPORT to disable compressed MusicXML file support
 * Making Verovio thread safe (@eNote-GmbH)
 * Switch from XSLT to Python for font generation (@eNote-GmbH)